[generate-compile_commands] Create a compilation database by default
This is useful for clang-based tools, in particular language servers like
clangd. I think it's a good idea to provide it by default.  I didn't find
a good way to detect a development setup, so I just made it the default and
disable it in scripts that are mostly for CI. Turned out quite crazy :(

If CMake would read environment variables like Make, then a developer could
use "export CMAKE_EXPORT_COMPILE_COMMANDS=ON" (once, for all projects),
but that doesn't work.